+++ text_explainability/data/sampling.py
@@ -2,7 +2,6 @@
 
 Todo:
     * Sample (informative?) subset from data
-    * Prototype sampling (k-medoids)
     * Refactor to make sampling base class
     * Add ability to perform MMD critic on a subset (e.g. single class)
 """
@@ -20,7 +19,89 @@ from text_explainability.data.weights import exponential_kernel
 from text_explainability.default import Readable
 
 
-class MMDCritic(Readable):
+class PrototypeSampler(Readable):
+    def __init__(self,
+                 instances: MemoryBucketProvider,
+                 embedder: Embedder = SentenceTransformer()):
+        """Generic class for sampling prototypes (representative samples) based on embedding distances.
+
+        Args:
+            instances (MemoryBucketProvider): Instances to select from (e.g. training set, all instance from class 0).
+            embedder (Embedder, optional): Method to embed instances (if the `.vector` property is not yet set). 
+                Defaults to SentenceTransformer().
+        """
+        self.embedder = embedder
+        self.instances = self.embedder(instances) if any(instances[i].vector is None for i in instances) \
+                         else instances
+
+    @property
+    def embedded(self) -> np.ndarray:
+        return np.stack(self.instances.bulk_get_vectors(list(self.instances))[-1])
+
+    def _select_from_provider(self, keys: Sequence[int]) -> Sequence[MemoryTextInstance]:
+        """Select instances from provider by keys."""
+        return [self.instances[i] for i in keys]
+
+    def prototypes(self, n: int = 5) -> Sequence[MemoryTextInstance]:
+        """Select `n` prototypes.
+
+        Args:
+            n (int, optional): Number of prototypes to select. Defaults to 5.
+
+        Returns:
+            Sequence[MemoryTextInstance]: List of prototype instances.
+        """
+        raise NotImplementedError('Implemented in subclasses')
+
+    def __call__(self, *args, **kwargs):
+        return self.prototypes(*args, **kwargs)
+
+
+class KMedoids(PrototypeSampler):
+    def __init__(self,
+                 instances: MemoryBucketProvider,
+                 embedder: Embedder = SentenceTransformer(),
+                 seed: int = 0):
+        """Sampling prototypes (representative samples) based on embedding distances using `k-Medoids`_.
+
+        Args:
+            instances (MemoryBucketProvider): Instances to select from (e.g. training set, all instance from class 0).
+            embedder (Embedder, optional): Method to embed instances (if the `.vector` property is not yet set). 
+                Defaults to SentenceTransformer().
+            seed (int, optional): Seed for reproducibility. Defaults to 0.
+
+        .. _k-Medoids:
+            https://scikit-learn-extra.readthedocs.io/en/stable/generated/sklearn_extra.cluster.KMedoids.html
+        """
+        super().__init__(instances, embedder)
+        self._seed = seed
+
+    def prototypes(self,
+                   n: int = 5,
+                   metric: Union[str, Callable] = 'cosine',
+                   **kwargs) -> Sequence[MemoryTextInstance]:
+        """Select `n` prototypes (most representative samples) using `k-Medoids`_.
+
+        Args:
+            n (int, optional): Number of prototypes to select. Defaults to 5.
+            metrics (Union[str, Callable], optional): Distance metric used to calculate medoids (e.g. 'cosine', 
+                'euclidean' or your own function). See `pairwise distances` for a full list. Defaults to 'cosine'.
+            **kwargs: Optional arguments passed to `k-Medoids`_ constructor.
+
+        Returns:
+            Sequence[MemoryTextInstance]: List of prototype instances.
+
+        .. _k-Medoids:
+            https://scikit-learn-extra.readthedocs.io/en/stable/generated/sklearn_extra.cluster.KMedoids.html
+        .. _pairwise distances:
+            https://scikit-learn.org/stable/modules/generated/sklearn.metrics.pairwise_distances.html
+        """
+        from sklearn_extra.cluster import KMedoids
+        kmedoids = KMedoids(n_clusters=n, metric=metric, random_state=self._seed, **kwargs).fit(self.embedded)
+        return self._select_from_provider(kmedoids.medoid_indices_)
+
+
+class MMDCritic(PrototypeSampler):
     def __init__(self,
                  instances: MemoryBucketProvider,
                  embedder: Embedder = SentenceTransformer(),
@@ -36,23 +117,16 @@ class MMDCritic(Readable):
         .. _MMD-critic:
             https://christophm.github.io/interpretable-ml-book/proto.html
         """
-        self.embedder = embedder
+        super().__init__(instances, embedder)
         self.kernel = kernel
-        self.instances = self.embedder(instances) if any(instances[i].vector is None for i in instances) \
-                         else instances
         self._calculate_kernel()
         self._prototypes = None
         self._criticisms = None
 
     def _calculate_kernel(self):
         """Calculate kernel `K` and column totals `colsum`."""
-        instances = np.stack(self.instances.bulk_get_vectors(list(self.instances))[-1])
-        self.K = self.kernel(instances, 1.0 / instances.shape[1])
-        self.colsum = np.sum(self.K, axis=0) / instances.shape[1]
-
-    def _select_from_provider(self, keys: Sequence[int]) -> Sequence[MemoryTextInstance]:
-        """Select instances from provider by keys."""
-        return [self.instances[i] for i in keys]
+        self.K = self.kernel(self.embedded, 1.0 / self.embedded.shape[1])
+        self.colsum = np.sum(self.K, axis=0) / self.embedded.shape[1]
 
     def prototypes(self, n: int = 5) -> Sequence[MemoryTextInstance]:
         """Select `n` prototypes (most representatitve instances), using `MMD-critic implementation`_.
@@ -187,13 +261,100 @@ class MMDCritic(Readable):
                 'criticisms': self.criticisms(n=n_criticisms, regularizer=regularizer)}
 
 
-class LabelwiseMMDCritic(Readable):
+class LabelwisePrototypeSampler(Readable):
+    def __init__(self,
+                 sampler: PrototypeSampler,
+                 instances: MemoryBucketProvider,
+                 labels: Union[Sequence[str], Sequence[int], LabelProvider],
+                 embedder: Embedder = SentenceTransformer(),
+                 **kwargs):
+        """Apply `PrototypeSampler()` for each label.
+
+        Args:
+            sampler (PrototypeSampler): Prototype sampler to construct (e.g. `KMedoids`, `MMDCritic`)
+            instances (MemoryBucketProvider): Instances to select from (e.g. training set, all instance from class 0).
+            labels (Union[Sequence[str], Sequence[int], LabelProvider]): Ground-truth or predicted labels, providing 
+                the groups (e.g. classes) in which to subdivide the instances.
+            embedder (Embedder, optional): Method to embed instances (if the `.vector` property is not yet set). 
+                Defaults to SentenceTransformer().
+            **kwargs: Additional arguments passed to `_setup_instances()` constructor.
+        """
+        self.sampler = sampler if isinstance(sampler, type) else self.sampler.__class__
+        if not isinstance(labels, LabelProvider):
+            labels = MemoryLabelProvider.from_tuples((id, frozenset({label}))
+                                                     for id, label in zip(list(instances), labels))
+        self.labels = labels
+        self.instances = instances
+        self._setup_instances(embedder, **kwargs)
+
+    def _setup_instances(self,
+                         embedder: Embedder,
+                         **kwargs):
+        """Setup a sampler for each label in `self.labels.labelset`.
+
+        Args:
+            embedder (Embedder): Method to embed instances (if the `.vector` property is not yet set). 
+                Defaults to SentenceTransformer().
+            **kwargs: Additional arguments passed to sampler constructor.
+        """
+        import copy
+
+        def select_by_label(label):
+            instances = copy.deepcopy(self.instances)
+            keys_to_keep = self.labels.get_instances_by_label(label)
+            instances._remove_from_bucket(frozenset(list(instances)).difference(keys_to_keep))
+            return instances
+
+        self._samplers = {label: self.sampler(instances=select_by_label(label),
+                                              embedder=embedder,
+                                              **kwargs)
+                          for label in self.labels.labelset}
+
+    def prototypes(self, n: int = 5) -> Dict[str, Sequence[MemoryTextInstance]]:
+        """Select `n` prototypes (most representatitve instances).
+
+        Args:
+            n (int, optional): Number of prototypes to select. Defaults to 5.
+
+        Returns:
+            Dict[str, Sequence[MemoryTextInstance]]: Dictionary with labels and corresponding list of prototypes.
+        """
+        return {label: sampler.prototypes(n=n)
+                for label, sampler in self._samplers.items()}
+
+
+class LabelwiseKMedoids(LabelwisePrototypeSampler):
+    def __init__(self,
+                 instances: MemoryBucketProvider,
+                 labels: Union[Sequence[str], Sequence[int], LabelProvider],
+                 embedder: Embedder = SentenceTransformer(),
+                 seed: int = 0):
+        """Select prototypes for each label based on embedding distances using `k-Medoids`_.
+
+        Args:
+            instances (MemoryBucketProvider): Instances to select from (e.g. training set, all instance from class 0).
+            labels (Union[Sequence[str], Sequence[int], LabelProvider]): Ground-truth or predicted labels, providing 
+                the groups (e.g. classes) in which to subdivide the instances.
+            embedder (Embedder, optional): Method to embed instances (if the `.vector` property is not yet set). 
+                Defaults to SentenceTransformer().
+            seed (int, optional): Seed for reproducibility. Defaults to 0.
+
+        .. _k-Medoids:
+            https://scikit-learn-extra.readthedocs.io/en/stable/generated/sklearn_extra.cluster.KMedoids.html
+        """
+        super().__init__(instances=instances,
+                         labels=labels,
+                         embedder=embedder,
+                         seed=seed)
+
+
+class LabelwiseMMDCritic(LabelwisePrototypeSampler):
     def __init__(self,
                  instances: MemoryBucketProvider,
                  labels: Union[Sequence[str], Sequence[int], LabelProvider],
                  embedder: Embedder = SentenceTransformer(),
                  kernel: Callable = exponential_kernel):
-        """Select  prototypes and criticisms for each label based on embedding distances using `MMD-Critic`_.
+        """Select prototypes and criticisms for each label based on embedding distances using `MMD-Critic`_.
 
         Args:
             instances (MemoryBucketProvider): Instances to select from (e.g. training set, all instance from class 0).
@@ -206,47 +367,10 @@ class LabelwiseMMDCritic(Readable):
         .. _MMD-critic:
             https://christophm.github.io/interpretable-ml-book/proto.html
         """
-        if not isinstance(labels, LabelProvider):
-            labels = MemoryLabelProvider.from_tuples((id, frozenset({label}))
-                                                     for id, label in zip(list(instances), labels))
-        self.labels = labels
-        self.instances = instances
-        self._setup_critics(embedder, kernel)
-
-    def _setup_critics(self,
-                       embedder: Embedder,
-                       kernel: Callable):
-        """Setup a `text_explainability.data.sampling.MMDCritic` for each label in `self.labels.labelset`.
-
-        Args:
-            embedder (Embedder): Method to embed instances (if the `.vector` property is not yet set). 
-                Defaults to SentenceTransformer().
-            kernel (Callable): Kernel to calculate distances.
-        """
-        import copy
-
-        def select_by_label(label):
-            instances = copy.deepcopy(self.instances)
-            keys_to_keep = self.labels.get_instances_by_label(label)
-            instances._remove_from_bucket(frozenset(list(instances)).difference(keys_to_keep))
-            return instances
-
-        self._critics = {label: MMDCritic(select_by_label(label),
-                                          embedder=embedder,
-                                          kernel=kernel)
-                         for label in self.labels.labelset}
-
-    def prototypes(self, n: int = 5) -> Dict[str, Sequence[MemoryTextInstance]]:
-        """Select `n` prototypes (most representatitve instances).
-
-        Args:
-            n (int, optional): Number of prototypes to select. Defaults to 5.
-
-        Returns:
-            Dict[str, Sequence[MemoryTextInstance]]: Dictionary with labels and corresponding list of prototypes.
-        """
-        return {label: critic.prototypes(n=n)
-                for label, critic in self._critics.items()}
+        super().__init__(instances=instances,
+                         labels=labels,
+                         embedder=embedder,
+                         kernel=kernel)
 
     def criticisms(self, n: int = 5, regularizer: Optional[str] = None) -> Dict[str, Sequence[MemoryTextInstance]]:
         """Select `n` criticisms (instances not well represented by prototypes).
@@ -262,8 +386,8 @@ class LabelwiseMMDCritic(Readable):
         Returns:
             Dict[str, Sequence[MemoryTextInstance]]: Dictionary with labels and corresponding list of criticisms.
         """
-        return {label: critic.criticisms(n=n, regularizer=regularizer)
-                for label, critic in self._critics.items()}
+        return {label: sampler.criticisms(n=n, regularizer=regularizer)
+                for label, sampler in self._samplers.items()}
 
     def __call__(self,
                  n_prototypes: int = 5,
@@ -281,5 +405,5 @@ class LabelwiseMMDCritic(Readable):
             Dict[str, Dict[str, Sequence[MemoryTextInstance]]]: Dictionary with labels and corresponding dictionary 
                 containing prototypes and criticisms.
         """
-        return {label: critic(n_prototypes=n_prototypes, n_criticisms=n_criticisms, regularizer=regularizer)
-                for label, critic in self._critics.items()}
+        return {label: sampler(n_prototypes=n_prototypes, n_criticisms=n_criticisms, regularizer=regularizer)
+                for label, sampler in self._samplers.items()}