+++ CHANGELOG.md
@@ -5,10 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Initial support for embeddings/vectors
+- Support for dimensionality reduction
 
 ## [0.3.8] - 2021-09-07
 ### Changed
 - Bugfix in including `locale/*.json` files during setup
+- Support for dimensionality reduction
 
 ## [0.3.7] - 2021-09-07
 ### Added